optimized cnn as much as i could
please change stats after using and push immediately, dont just run the test file and leave it as such.
Dont modify ANYTHING else in the cnn.py file
It works amazing for almost all confirmed and deaths, but for some like botswana it gives logically correct outputs but relative error is high due to erratic data and low numbers.
It doesnt perform too well for recovered, so if anyone else is trying a model, first try to optimize for recovered